no input norm and beta float

+++ swwae/run.py
@@ -31,7 +31,7 @@ parser.add_argument("--num_it", type=int, default=300000,
                     help='iteration number')
 parser.add_argument("--net_archi", default='conv',
                     help='networks architecture [mlp/conv]')
-parser.add_argument("--beta", type=int, default=10.,
+parser.add_argument("--beta", type=float, default=10.,
                     help='Latent reg weight setup')
 parser.add_argument("--sigma_pen", action='store_true', default=False,
                     help='penalization of Sigma_q')
